6.0.0-rc-0

+++ src/NEventStore.Persistence.Sql/CommitExtensions.cs
@@ -47,9 +47,9 @@ namespace NEventStore.Persistence.Sql
             return record[CommitSequenceIndex].ToInt();
         }
 
-        public static int CheckpointNumber(this IDataRecord record)
+        public static long CheckpointNumber(this IDataRecord record)
         {
-            return record[CheckpointIndex].ToInt();
+            return record[CheckpointIndex].ToLong();
         }
 
         public static T Deserialize<T>(this ISerialize serializer, IDataRecord record, int index)
